hide client socket to mitigate unsafe permissions change
If the libchrony application is running under root and it binds its Unix
domain socket, it needs to change the socket permissions in order for
chronyd running without root privileges to be able to send a response to
the socket.

There is a race condition between the bind() and chmod() calls. If an
attacker was able to execute arbitrary code in the chronyd process, it
might be able to wait for the application to be executed under root,
replace the socket with a symlink between the two calls, and cause the
privileged process to change permissions of something else, possibly
leading to a privilege escalation.

There doesn't seem to be a safe and portable way to change the socket
permissions directly. Changing the process umask is problematic with
threads.

Hide the socket in two levels of subdirectories (the lower one having
a long random name not visible to the chronyd process) to make the
socket path unpredictable, and force the bind() or chmod() call to fail
if the visible upper directory is replaced.

To avoid additional complexity, at least for now, drop the ability to
replace existing sockets left by uncleanly terminated libchrony
applications.

(The same change was implemented in chronyc in chrony-4.8.)